pyproject.toml
[project]
name = "uom"
version = "0.1.0"
description = "CLI media library manager — scan, tag, search, organise, deduplicate."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiosqlite>=0.22.1",
    "click>=8.3.1",
    "fastapi[standard]>=0.115",
    "numpy>=1.24",
    "peewee>=3.19.0",
    "peewee-aio>=2.2.2",
    "peewee-async>=1.2.1",
    "pillow>=10.0",
    "pillow-heif>=0.18",
    "rawpy>=0.26.1",
]

[project.optional-dependencies]
clip = ["open-clip-torch>=2.24", "torch>=2.0", "pillow>=10.0"]

[project.scripts]
uom = "uom.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/uom"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true

[dependency-groups]
dev = ["pytest>=9.0.2"]
